cleanup file names
