+++ flightbooker/src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import HomeBookingCard from "../components/Home/HomeBookingCardv3";
+import HomeBookingCard from "../components/Home/HomeBookingCard";
 import { HomeBookingCardContextProvider } from "../providers/HomeBookingCardContextProvider";
 
 const Home: React.FC = () => {
